Replace gmp with num

+++ src/text_query.rs
@@ -488,7 +488,7 @@ fn parse_function(iter: &mut Iter<'_>, func: Function) -> Expr {
     Expr::Call(func, args)
 }
 
-fn parse_radix(num: &str, base: u8, description: &str) -> Expr {
+fn parse_radix(num: &str, base: u32, description: &str) -> Expr {
     BigInt::from_str_radix(num, base)
         .map(|x| BigRat::ratio(&x, &BigInt::one()))
         .map(Num::Rational)
